# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MedCenter Back",
    "description": "API for MedCenter",
    "version": "1.0.0"
  },
  "host": "localhost:8088",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/clinic/create": {
      "post": {
        "tags": [
          "Clinic"
        ],
        "summary": "Созданий нового типа услуг",
        "description": "Добавление нового типа услуг",
        "consumes": [
          "name"
        ],
        "produces": [
          "name"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Название типа"
          },
          {
            "name": "icon",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Иконка типа"
          }
        ],
        "responses": {
          "201": {
            "description": "Новый тип услуг был зарегистирован",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось создать новый тип услуг",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/clinic/list": {
      "get": {
        "tags": [
          "Clinic"
        ],
        "summary": "Получение всех типов услуг клиники",
        "description": "Получение списка всех типов услуг клиники",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "description": "Фильтры для поиска определенных услуг",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "test"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список всех услуг поликлиники",
            "schema": {
              "$ref": "#/definitions/ClinicList"
            }
          }
        }
      }
    },
    "/consult/create": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "Создание консультации",
        "description": "Созданией консультации",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Данные для создания консультации",
            "schema": {
              "$ref": "#/definitions/Consult"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Запрос на консультацию был отправлен",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось отправить запрос на консультацию",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/consult/list": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "Получение всех заявок на консультации",
        "description": "Получение списка всех заявок на консультации",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "description": "Фильтры для поиска консультаций",
            "schema": {
              "$ref": "#/definitions/Consult"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список всех заявок",
            "schema": {
              "$ref": "#/definitions/ConsultList"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/review/create": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Создание консультации",
        "description": "Созданией консультации",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Данные для создания консультации",
            "schema": {
              "$ref": "#/definitions/Consult"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Отзыв был создан",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось создать отзыв",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/review/list": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Получение всех отзывов по фильтрам",
        "description": "Получение всех отзывов по фильтрам",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "description": "Фильтры для поиска отзывов",
            "schema": {
              "$ref": "#/definitions/ReviewList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список всех отзывов",
            "schema": {
              "$ref": "#/definitions/ReviewFilters"
            }
          }
        }
      }
    },
    "/doctor/list": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Получение всех докторов по фильтрам",
        "description": "Получение всех докторов по фильтрам",
        "consumes": [
          "body"
        ],
        "produces": [
          "body"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": false,
            "description": "Фильтры для поиска доктора",
            "schema": {
              "$ref": "#/definitions/ReviewList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список всех отзывов",
            "schema": {
              "$ref": "#/definitions/DoctorFilters"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Message": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Сообщение"
        }
      }
    },
...
    "ClinicList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Тип услуги"
              },
              "icon": {
                "type": "string",
                "example": "/путь/к изображению"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
    "Consult": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "Андрей Голубцов"
        },
        "сomplaints": {
          "type": "string",
          "example": "Колики в животике"
        },
        "phone": {
          "type": "string",
          "example": "8 988 982 23 23"
        }
      }
    },
    "ConsultList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "example": "Андрей Голубцов"
              },
              "сomplaints": {
                "type": "string",
                "example": "Колики в животике"
              },
              "phone": {
                "type": "string",
                "example": "8 988 982 23 23"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
...
    "DoctorFilters": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "name": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Геннадий Васильевич ..."
                }
              }
            },
            "specialization": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "Терапия"
                }
              }
            }
          }
        }
      }
    },
...
    "ReviewFilters": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "отзыв..."
        },
        "rating": {
          "type": "number",
          "example": 4
        }
      }
    },
    "ReviewList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "отзыв"
              },
              "rating": {
                "type": "number",
                "example": 4
              },
              "userId": {
                "type": "number",
                "example": 1
              },
              "doctorId": {
                "type": "number",
                "example": 1
              },
              "createdAt": {
                "type": "string",
                "example": "2025-10-11"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    }
  }
}